requirements.txt
# AWS SDK
boto3>=1.34.0

# Pydantic AI
pydantic-ai>=0.0.14

# Supabase Python Client
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Text processing
langchain-text-splitters>=0.3.0
langchain-community>=0.3.0

# AWS Lambda (for deployment)
aws-lambda-logging>=0.1.1

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# CLI
prompt-toolkit>=3.0.0
